@@ -1,58 +1,23 @@
 {
-  "homepage": "http://hs-credit-dev.github.io/frontend",
-  "name": "hscredit",
+  "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
-    "@mui/core": "^5.0.0-alpha.54",
-    "@mui/icons-material": "^5.2.5",
-    "@mui/material": "^5.2.5",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/amqplib": "^0.8.2",
-    "axios": "^0.21.4",
-    "bcrypt": "^5.0.1",
-    "bootstrap": "^5.1.3",
-    "classnames": "^2.3.1",
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "gh-pages": "^3.2.3",
-    "jsonwebtoken": "^8.5.1",
-    "jwt-decode": "^3.1.2",
-    "mdbootstrap": "^4.20.0",
-    "mdbreact": "^5.2.0",
-    "mongoose": "^5.12.14",
-    "nodemon": "^2.0.13",
-    "passport": "^0.4.1",
-    "passport-jwt": "^4.0.0",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.0.3",
-    "react-bootstrap-carousel": "^4.1.1",
-    "react-dom": "^17.0.2",
-    "react-multi-carousel": "^2.6.5",
-    "react-player": "^2.9.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-swipeable-views": "^0.14.0",
-    "react-update": "^0.4.4",
-    "reactstrap": "^8.10.1",
-    "rsuite": "^4.10.2",
-    "tailwind": "^4.0.0",
-    "validator": "^13.6.0",
-    "web-vitals": "^1.1.2"
+    "@aws-amplify/ui-react": "^3.2.1",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "aws-amplify": "^4.3.28",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d build"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
